requirements.txt
google-generativeai
vertexai
diffusers
torch
transformers
Pillow
SpeechRecognition
